Dev: (Mod store) Move ongoing operations and store selector to sidebar, split stuff into modules
